Fix build errors in latest Android Studio
The first time I opened the project, I got greeted with a Java version not supported error. After updating gradle wrapper to the latest version, I then got a "flavor dimension" error. A search on stackoverflow suggested that upgrading AGP version will solve the problem, so I did. In process of doing so, I had to move the "package" from the manifest file to a "namespace" property in the build.gradle. Now I'm finally able to build the project.